Refactor all changes as requested
