+++ integration-tests/sceptre-project/config/9/B.yaml
@@ -1,4 +1,4 @@
 template_path: dependencies/dependent_template_local_export.json
 region: eu-west-1
 parameters:
-  DependentStackName: !stack_output_external "{project_code}-9-A::StackName default::eu-west-1"
+  DependentStackName: !stack_output_external "sceptre-integration-tests-452379e268b711ecaf07acde48001122-9-A::StackName default::eu-west-1"
