+++ integration-tests/sceptre-project/config/drift-group/A.yaml
