+++ integration-tests/sceptre-project/config/drift-group/B.yaml
