+++ integration-tests/sceptre-project/config/drift-single/A.yaml
