+++ integration-tests/sceptre-project/templates/loggroup.yaml
@@ -1,0 +1,11 @@
+Resources:
+  LogGroup:
+    Type: AWS::Logs::LogGroup
+    Properties:
+      RetentionInDays: 1
+
+Outputs:
+  LogGroup:
+    Value: !Ref LogGroup
+    Export:
+      Name: !Sub "${AWS::StackName}-LogGroup"
